Debug reset problem
SUCCESS!

Had to add while(1) or else (I'm guessing) the instruction register
walks off to infinity, and then when the interrupt returns, it's
set to an illegal value, resulting in triple fault.

Also, when IDT was setting value 0, it would reset the CPU (no fault).
I think this is because the keyboard was triggering interrupt 0, which
seems to be the divide-by-zero exception.