Add debug mode - `make debug`
Also fixed a horredous error - forgot `ret` in ioport_out.
That explains the corrupted video data!
Every time we did an ioport_out, it kept running, right into
print_char_with_asm, using whatever values happened to be in
the registers at the time.

I'm surprised it even worked with it like this!